Moving dask test to test_008
